Added books_mysql spider that saves data to MySql database

+++ books_crawler/spiders/books_only_scrapy.py
@@ -1,6 +1,8 @@
 import os
 import glob
 import re
+import csv
+import mysql.connector
 from scrapy import Spider
 from scrapy.http import Request
 from scrapy.loader import ItemLoader
@@ -16,7 +18,9 @@ class BooksOnlyScrapySpider(Spider):
     
     def __init__(self, category=HOME_PAGE):
         # delete previous exported csv
-        os.remove(get_csv_filename(category))
+        if os.path.isfile(get_csv_filename(category)):
+            os.remove(get_csv_filename(category))
+
         self.start_urls = [category]
         self.category = category
 
@@ -26,14 +30,15 @@ class BooksOnlyScrapySpider(Spider):
             absolute_url = response.urljoin(book)
             yield Request(absolute_url, callback=self.parse_book)
 
-        next_page_url = response.xpath('//a[text()="next"]/@href').extract_first()    
-        absolute_next_page_url = response.urljoin(next_page_url)
-        yield Request(absolute_next_page_url)
+        # - Process next page
+        # next_page_url = response.xpath('//a[text()="next"]/@href').extract_first()    
+        # absolute_next_page_url = response.urljoin(next_page_url)
+        # yield Request(absolute_next_page_url)
 
     def parse_book(self, response):
         l = ItemLoader(item=BooksCrawlerItem(), response=response)
         title = response.xpath('//article[@class="product_page"]//h1/text()').extract_first()
-        l.add_value('title', title)
+        l.add_value('title', title) 
         price = response.xpath('//p[@class="price_color"]/text()').extract_first()
         l.add_value('price', price)
         description = response.xpath('//*[@id="product_description"]/following-sibling::p/text()').extract_first()          
@@ -79,15 +84,6 @@ class BooksOnlyScrapySpider(Spider):
         return l.load_item()
         
     def close(self, reason):
-        # category_name = ''
-        # if self.category != self.HOME_PAGE:
-        #     cat_regex_result = re.compile("\/books\/(.+)_\d+\/").search(self.category)
-        #     if cat_regex_result is not None:
-        #         category_name = cat_regex_result.group(1)
-        # csv_name = 'books.csv'
-        # if category_name != '':
-        #     csv_name = 'books-{0}.csv'.format(category_name)  
-          
         csv_file = max(glob.iglob('*.csv'), key=os.path.getctime)
         os.rename(csv_file, get_csv_filename(self.category))
         
@@ -96,7 +92,7 @@ def get_csv_filename(category):
     cat_regex_result = re.compile("\/books\/(.+)_\d+\/").search(category)
     if cat_regex_result is not None:
         category_name = cat_regex_result.group(1)
-        
+
     csv_name = 'books.csv'
     if category_name != '':
         csv_name = 'books-{0}.csv'.format(category_name)
